// src/api/generated/model/orderListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { OrderResponse } from "./orderResponse.ts";

export interface OrderListResponse {
	items?: OrderResponse[];
}